Create Use Recursion to Create a Range of Numbers.py
The function return an array of integers which begins with a number represented by the startNum parameter and ends with a number represented by the endNum parameter. The starting number will always be less than or equal to the ending number. The function use recursion by calling itself and not use loops of any kind. It also work for cases where both startNum and endNum are the same.